// src/frontend/src/backend-shim.d.ts
export interface Task {
  id: bigint;
  title: string;
  subjectName: string;
  createdAt: bigint;
  completed: boolean;
  dueDate: string;
  priority: string;
}